Shared module for common argument types
